move presentaion completion to presentation
